+++ src/main/java/de/samply/db/crud/TeilerDbService.java
@@ -2,7 +2,10 @@ package de.samply.db.crud;
 
 import de.samply.db.model.Query;
 import de.samply.db.repository.QueryRepository;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -21,5 +24,14 @@ public class TeilerDbService {
   public String saveQueryAndGetQueryId (Query query){
     return queryRepository.save(query).getId().toString();
   }
+
+  public List<Query> fetchAllQueries(){
+    return queryRepository.findAll();
+  }
+
+  public List<Query> fetchAllQueries(int page, int pageSize){
+    return queryRepository.findAll(PageRequest.of(page, pageSize)).stream().toList();
+  }
+
 
 }